requirements.txt
psycopg2==2.9.1                # PostgreSQL adapter for Python
gensim==4.1.2                  # LDA and other NLP models
sentence-transformers==2.2.0    # Pre-trained transformer models for embeddings
faiss-cpu==1.7.2               # FAISS library for similarity search
vaderSentiment==3.3.2          # VADER sentiment analysis
geopy==2.2.0                   # Geocoding library (Nominatim or other services)
nltk==3.6.7                    # Natural Language Toolkit for tokenization, stopwords, etc.
transformers==4.11.3           # Hugging Face transformers for NER
torch==1.10.0                  # PyTorch (needed for Hugging Face models)
numpy==1.21.2                  # Used in several models and FAISS

# Optional libraries that might be useful
pandas==1.3.3                  # For data manipulation and querying (optional)
requests==2.26.0               # If you need to make any API requests (optional)
